+++ front/src/common/fragments.ts
@@ -391,6 +391,15 @@ export const detailFormFragment = gql`
       }
     }
     quantityGrouped
+    intermediaries {
+      name
+      siret
+      vatNumber
+      phone
+      contact
+      mail
+      address
+    }
   }
   ${transporterFormFragment}
 `;